feat(simple buy): code cleanup moved waitForUserId to be part of profile saga
